refactor(category): add everything category related inside charts/cards/crud sub-folders
