+++ FinEdgeFrontend/src/Components/CategoriesPage/Cards/CategoryData.tsx
